Create README.md
This is a simple practice program I wrote to revive my coding experience. The program is given a list of possible new pets, randomly samples from that list a user-defined number of times, and replicates the sample a user-defined number of times. The results are visualized in a histogram showing how many times each new pet is chosen. Statistics are not implemented in the program, but would be interesting to include. The user can play with the number of samples and replicates to watch how those parameters influence the outcome of the graph. I was particularly interested in finding the parameters at which point the graph started to represent a 'true random sample,' in which all options have an equal probability of being chosen, which would present itself in the histogram as all 'pets' being chosen approximately the same number of times.